Schedule the release PR to arrive weekly (#7325)
This adds a workflow trigger that goes off every Thursday. This way we
are reminded to release weekly.

If the version bump PR is already open because it hasn't yet been merged
from the previous week, then this workflow will bomb out when it tries
to push the branch, since the branch name is idempotent.