The different sections of the metadata go through the LinkAndMentionSanitizer twice. (#7190)
Once per sub section and once as the complete message.

In case of source providers that support HTML that is not a problem.
The Markdown is converted to HTML and after that left alone.

In case of a source provider that does not support HTML there is a problem.
The links in the markdown are converted from github.com to redirect.github.com.
And in the second pass the links are converted from redirect.github.com to redirect.redirect.github.com.