Turn `composer` version into a metric (#7323)
We'd like to know how many jobs are running using `composer 1` vs
`composer 2`, so this exposes it as a metric.

WARNING: This is a bit of a hack, because it doesn't actually report the
_user's_ range of supported versions, it only reports the version that
_dependabot_ chose to use. We'd much rather know the range of versions
in use. However, doing that will take a bit more work, and I've already
got this code sitting around and it's helpful to at least know something
rather than nothing at all. This should also give us a ballpark figure
on how often we're using `composer` `v1` vs `v2`, which will let us know
when we can drop `v1`.

Additionally, parsing the `composer.json` file to derive the list of
allowed versions will be simpler once we don't have to worry about `v1`,
so if this gets us far enough to drop `v1`, then great!

Long term, this can be further refined to report `composer` versions
with more granularity, for example `2.3`, `2.4`, `2.5`, etc.