Remove `3.6` guard when setting `poetry config experimental.system-git-client` (#7614)
We no longer support Python 3.6, so this now always needs to happen.

I also bumped the minimum version of `poetry` up to `1.2` as that's when
the flag was added. It shouldn't matter, as python will default to the
highest version allowed by that version of python, and now that we only
support `3.7` onwards that will be the upper bound.