Use a better strategy to parse Poetry dependency types
Instead of parsing the lockfile directly, use `poetry show --only main`,
and consider the dependencies there as "production", and the others as
"development".

This has the advantage of not having to parse the lockfiles directly,
and as a result fixes an issue where we try to parse a "category"
attribute in the lockfile which has been removed in the latest lockfile
format.

Ideally `poetry show` would support JSON output, but for now this should
be good enough.

Co-authored-by: Phillip Verheyden <pverheyden@gmail.com>