Ignore repo access issues while parsing actions dependencies
The functionality that needs accessing the dependency source is not
critical, so skip it if the repo is not accessible. The end result of
this should be that we still parse the dependency, and it will be
properly ignore before checking for updates for it. If it's a non
ignored dependency, we should be getting another
`git_dependencies_not_reachable` error when checking for updates.