Use a URL that won't respond (#11986)
`https://composer.dependabot.com` no longer has a DNS record now that
Dependabot was acquired by GitHub. And we can't just stub the URL
because the call is happening within the `composer` subprocess, not in
our ruby process.

This helpful StackOverflow pointed out that example.com is owned by IANA
so should be safe, and port 81 is never expected to respond:
https://stackoverflow.com/questions/100841/artificially-create-a-connection-timeout-error

Alternatively, we could use a non-routable IP but that may return
different errors depending on the topology of the LAN where it's running.